get to work

+++ clientserver/messagehandler.cc
@@ -32,7 +32,7 @@ string MessageHandler::readString(){
 string MessageHandler::readString(int n){
 	string s;
 	char ch;
-	for (size_t i = 0; i != n;++i) {
+	for (int i = 0; i != n;++i) {
 		ch = conn.read();
 		s += ch;
 	}
